% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Stack_maps.R
\name{paramo}
\alias{paramo}
\title{Final stack of maps
cc chars id to stack
ntrees number of trees to stack
dirW directory for zip file}
\usage{
paramo(cc, ntrees = 10, dirW = c(""))
}
\description{
Final stack of maps
cc chars id to stack
ntrees number of trees to stack
dirW directory for zip file
}
