% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{prepareMaps}
\alias{prepareMaps}
\title{Function for processing revbayes stochastic maps}
\usage{
prepareMaps(td, dirR, dirW, discretization_level = 100, start_tree = 1,
  end_tree = 2)
}
\description{
Function for processing revbayes stochastic maps
}
